Ensure file server asset handling points to asset directory
Fixes bug when used in an environment in which asset_host is configured
to a remote CDN; the rack file server cannot serve a file off a remote
host. Using `compute_asset_host` will look up the local file path. When
the method is not available, we pass in a host override.